BizLedger/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
whisper-openai==20231117
torch==2.1.1
torchaudio==2.1.1
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1